clinical_annotator/requirements.txt
Django==5.2
pandas==1.5.3
Requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
